+++ plugin/verticator/plugin-src.js
@@ -1,5 +1,12 @@
 const Plugin = () => {
 
+	const lightClass = `has-light-background`;
+	const darkClass = `has-dark-background`;
+	const themeColorVar = `--c-theme-color`;
+	const vertiColorVar = `--v-color`;
+	const forceColorVar = `--v-forcecolor`;
+	const activeclass = 'active';
+
 	const loadStyle = function(url, type, callback) {
 		let head = document.querySelector('head');
 		let style;
@@ -7,21 +14,47 @@ const Plugin = () => {
 		style.rel = 'stylesheet';
 		style.href = url;
 
-		let finish = function () {
-		  if (typeof callback === 'function') {
-			callback.call();
-			callback = null;
-		  }
-		};
-	
+		let finish = function () { if (typeof callback === 'function') { callback.call(); callback = null }};
 		style.onload = finish;
 	
-		style.onreadystatechange = function () {
-		  if (this.readyState === 'loaded') {
-			finish();
-		  }
-		};
+		style.onreadystatechange = function () { if (this.readyState === 'loaded') { finish() }};
 		head.appendChild(style);
+	}
+
+	const findThemeColors = function(parent, tag) {
+		const themeColors = {};
+		const colorSection = document.createElement("section");
+		const sectionChild = document.createElement(tag);
+		parent.getElementsByClassName('slides')[0].appendChild(colorSection).appendChild(sectionChild);
+		themeColors.regular = getComputedStyle(sectionChild).getPropertyValue('color');
+		colorSection.classList.add(lightClass);
+		themeColors.inverse = getComputedStyle(sectionChild).getPropertyValue('color');
+		if (themeColors.regular == themeColors.inverse) {
+			themeColors.theme = "light";
+			colorSection.classList.remove(lightClass);
+			colorSection.classList.add(darkClass);
+			themeColors.inverse = getComputedStyle(sectionChild).getPropertyValue('color');
+		} else {themeColors.theme = "dark";}
+		colorSection.remove();
+		return themeColors;
+	}
+
+	const swapColors = ( needToSwap, options, colors, revealElement, theVerticator) => {
+		if (needToSwap) {
+			revealElement.style.setProperty(themeColorVar, colors.themeinverse);
+			if (options.inversecolor || options.oppositecolor) {
+				theVerticator.style.setProperty(vertiColorVar, colors.verticatorinverse);
+			} else {
+				theVerticator.style.removeProperty(vertiColorVar);
+			}
+		} else {
+			revealElement.style.setProperty(themeColorVar, colors.themeregular);
+			if (options.color) {
+				theVerticator.style.setProperty(vertiColorVar, colors.verticatorregular);
+			} else {
+				theVerticator.style.removeProperty(vertiColorVar);
+			}
+		}
 	}
 
 	const verticate = function(deck, options) {
@@ -30,7 +63,6 @@ const Plugin = () => {
 		userScale = (userScale > 2) ? 2 : (userScale < 0.5) ? 0.5 : userScale;
 
 		let revealElement = deck.getRevealElement();
-
 		let theVerticator = revealElement.querySelector('ul.verticator');
 
 		if (!theVerticator) {
@@ -49,6 +81,99 @@ const Plugin = () => {
 		let totalScale = revealScale > 1 ? revealScale * userScale : userScale;
 		theVerticator.style.setProperty('--verticator-scale', totalScale.toFixed(2));
 
+		const colors = {};
+		const themeColors = findThemeColors(revealElement, options.themetag ? options.themetag : 'section');
+
+		colors.theme = themeColors.theme;
+		colors.themeregular = themeColors.regular;
+		colors.themeinverse = themeColors.inverse;
+
+		colors.verticatorregular = options.color ? options.color : themeColors.regular;
+		colors.verticatorinverse = options.inversecolor ? options.inversecolor : options.oppositecolor ? options.oppositecolor : themeColors.inverse;
+
+		if (options.debug) {
+			console.log(`Theme regular color is: "${colors.themeregular}"`);
+			console.log(`Theme inverse color is: "${colors.themeinverse}"`);
+			if (options.color) {
+				console.log(`Verticator regular color is: "${colors.verticatorregular}"`);
+			}
+			if (options.inversecolor || options.oppositecolor) {
+				console.log(`Verticator inverse color is: "${colors.verticatorinverse}"`)
+			}
+		}
+
+		revealElement.style.setProperty(themeColorVar, colors.themeregular);
+		if (options.color) {
+			theVerticator.style.setProperty(vertiColorVar, colors.verticatorregular);
+		}
+
+		let sectionState = {};
+		sectionState.dark = revealElement.classList.contains(darkClass);
+		sectionState.light = revealElement.classList.contains(lightClass);
+		sectionState.rtl = revealElement.classList.contains('rtl');
+
+		const revealElementObserver = new MutationObserver(function(mutations) {
+			mutations.forEach(function(mutation) {
+				const { target } = mutation;
+
+				if (mutation.attributeName === 'class') {
+
+					let hasLightBg = mutation.target.classList.contains(lightClass);
+					let hasDarkBg = mutation.target.classList.contains(darkClass);
+					let leftAligned = mutation.target.classList.contains('rtl');
+
+					if (sectionState.dark !== hasDarkBg || sectionState.light !== hasLightBg) {
+						if (colors.theme == "dark" && sectionState.light !== hasLightBg) {
+							sectionState.light = hasLightBg;
+							swapColors(hasLightBg, options, colors, revealElement, theVerticator)
+						}
+						if (colors.theme == "light" && sectionState.dark !== hasDarkBg) {
+							sectionState.dark = hasDarkBg;
+							swapColors(hasDarkBg, options, colors, revealElement, theVerticator)
+						}
+					}
+					if (options.position == "auto") {
+						if (sectionState.rtl !== leftAligned) {
+							sectionState.rtl = leftAligned;
+							if (leftAligned) {
+								theVerticator.style.left = options.offset;
+								theVerticator.style.right = "auto";
+								theVerticator.classList.add('left');
+							} else {
+								theVerticator.style.right = options.offset;
+								theVerticator.style.left = "auto";
+								theVerticator.classList.remove('left');
+							}
+						}
+					}
+
+				}
+			});
+		});
+
+		revealElementObserver.observe(revealElement, {attributes: true,  attributeFilter : ['class']});
+
+		deck.on('slidechanged', event => {
+			if (event.currentSlide.dataset.verticator) {
+				if (event.currentSlide.dataset.verticator == "regular") {
+					theVerticator.style.setProperty(forceColorVar, colors.verticatorregular);
+					if (options.debug) {
+						console.log(`Verticator forced to: "${colors.verticatorregular}"`);
+					}
+				} else if (event.currentSlide.dataset.verticator == "inverse") {
+					theVerticator.style.setProperty(forceColorVar, colors.verticatorinverse);
+					if (options.debug) {
+						console.log(`Verticator forced to: "${colors.verticatorinverse}"`);
+					}
+				} else {
+					theVerticator.style.setProperty(forceColorVar, deck.getCurrentSlide().dataset.verticator);
+					if (options.debug) {
+						console.log(`Verticator forced to: "${deck.getCurrentSlide().dataset.verticator}"`);
+					}
+				}
+			} else {theVerticator.style.removeProperty(forceColorVar);}
+		});
+
 		deck.on('resize', event => {
 			revealScale = event.scale;
 			totalScale = revealScale > 1 ? revealScale * userScale : userScale;
@@ -58,16 +183,16 @@ const Plugin = () => {
 		if (options.offset != '3vmin') {
 			theVerticator.style.right = options.offset;
 		}
-		if (options.position == 'left') {
-			theVerticator.classList.add('left');
-			theVerticator.style.right = 'auto';
-			theVerticator.style.left = options.offset;
-		}
-		if (options.position != 'left' && options.position != 'right') {
-			options.position = 'right'
-		}
 
-		let activeclass = 'active';
+		theVerticator.style.right = options.offset;
+		if (options.position !== "auto") {
+			let opposite = options.position == "left" ? "right" : "left";
+			theVerticator.style[options.position] = options.offset;
+			theVerticator.style[opposite] = "auto";
+			if (options.position == "left" && options.tooltip !== false) {
+				theVerticator.classList.add('left');
+			}
+		}
 
 		const selectionArray = function(container, selectors) {
 			let selections = container.querySelectorAll(selectors);
@@ -76,12 +201,9 @@ const Plugin = () => {
 		};
 
 		const clickBullet = function(event) {
-			if ((event.target)
-				.matches('.verticator li a')) {
-				let currIndexh = (deck.getIndices())
-					.h;
-				let currIndexf = (deck.getIndices())
-					.v;
+			if ((event.target).matches('.verticator li a')) {
+				let currIndexh = (deck.getIndices()).h;
+				let currIndexf = (deck.getIndices()).v;
 				let i = getNodeindex(event.target.parentNode);
 				event.preventDefault();
 				deck.slide(currIndexh, i, currIndexf);
@@ -91,62 +213,23 @@ const Plugin = () => {
 		const activateBullet = function(event) {
 
 			let listItems = selectionArray(theVerticator, 'li');
-			let hasDarkBackground = false;
-			let hasLightBackground = false;
-
-			if (revealElement.classList.contains('has-dark-background')) {
-				hasDarkBackground = true;
-			}
-			if (revealElement.classList.contains('has-light-background')) {
-				hasLightBackground = true;
-			}
-
-			if (event.currentSlide.dataset.state) {
-				let currentState = event.currentSlide.dataset.state.split(' ');	
-				if (currentState.includes("has-dark-background")) {
-					hasDarkBackground = true;
-				}
-				if (currentState.includes("has-light-background")) {
-					hasLightBackground = true;
-				}
-			}
-
-			if (hasDarkBackground) {
-				theVerticator.style.color = options.oppositecolor;
-				theVerticator.style.setProperty('--bullet-maincolor', options.oppositecolor);
-			} else {
-				theVerticator.style.color = options.color;
-				theVerticator.style.setProperty('--bullet-maincolor', options.color);
-			}
-
-			if (options.darktheme) {
-				if (hasLightBackground) {
-					theVerticator.style.color = options.oppositecolor;
-					theVerticator.style.setProperty('--bullet-maincolor', options.oppositecolor);
-				} else {
-					theVerticator.style.color = options.color;
-					theVerticator.style.setProperty('--bullet-maincolor', options.color);
-				}
-			}
 
 			var bestMatch = options.indexbase - 1;
 
 			listItems.forEach(function(listItem, i) {
-				if (parseInt(listItem.getAttribute("data-index")) <= event.indexv + options.indexbase) {
+				if (parseInt(listItem.dataset.index) <= event.indexv + options.indexbase) {
 					bestMatch = i;
 				}
-
 				listItem.classList.remove(activeclass);
 			});
 
 			if (bestMatch >= 0) {
 				listItems[bestMatch].classList.add(activeclass);
 			}
-
 		};
 
 		const ttName = function(element) {
-			if (element.getAttribute("data-verticator-tooltip") && (element.getAttribute("data-verticator-tooltip") == "none" || element.getAttribute("data-verticator-tooltip") == "false") || element.classList.contains('no-verticator-tooltip')) {
+			if (element.dataset.verticatorTooltip && (element.dataset.verticatorTooltip == "none" || element.dataset.verticatorTooltip == "false") || element.classList.contains('no-verticator-tooltip')) {
 				return
 			} else if (options.tooltip != "auto" && element.getAttribute(`${options.tooltip}`)) {
 				return element.getAttribute(`${options.tooltip}`)
@@ -158,8 +241,7 @@ const Plugin = () => {
 				}
 				for (const slctr of ["h1", "h2", "h3", "h4"]) {
 					if (element.querySelector(slctr)) {
-						return element.querySelector(slctr)
-							.textContent;
+						return element.querySelector(slctr).textContent;
 					}
 				}
 			} else return false
@@ -167,8 +249,8 @@ const Plugin = () => {
 
 		const createBullets = function(event, sections) {
 
-			theVerticator.style.color = options.color;
 			theVerticator.classList.remove('visible');
+
 			let listHtml = '';
 
 			sections.forEach(function(section) {
@@ -180,11 +262,12 @@ const Plugin = () => {
 				listHtml += `<li data-index="${i + options.indexbase}"><a ${options.clickable ? link : ''}${dataname}></a>${tooltip}</li>`;
 			});
 
+			theVerticator.innerHTML = listHtml;
+			activateBullet(event);
 			setTimeout(function() {
-				theVerticator.innerHTML = listHtml;
-				activateBullet(event);
 				theVerticator.classList.add('visible');
-			}, 200);
+			}, 300);
+
 		}
 
 		const slideAppear = function(event) {
@@ -222,17 +305,14 @@ const Plugin = () => {
 				} else {
 					createBullets(event, sections);
 				}
-
-				setTimeout(function() {
-					activateBullet(event);
-				}, 150);
+				activateBullet(event);
 			}
 		};
 
 		if (theVerticator) {
-			deck.on('slidechanged', event => {
-				slideAppear(event)
-			});
+			const eventnames = ['ready', 'slidechanged'];
+			eventnames.forEach( (eventname) => deck.on( eventname, event => { slideAppear(event) } ) )
+
 			if ((deck.getConfig())
 				.embedded) {
 				deck.on('click', event => {
@@ -240,7 +320,6 @@ const Plugin = () => {
 				});
 			}
 		}
-
 	};
 
 	const init = function(deck) {
@@ -248,12 +327,12 @@ const Plugin = () => {
 		let es5Filename = "verticator.js"
 
 		let defaultOptions = {
-			darktheme: false,
-			color: 'black',
-			oppositecolor: 'white',
+			themetag: 'h1',
+			color: '',
+			inversecolor: '',
 			skipuncounted: false,
 			clickable: true,
-			position: 'right',
+			position: 'auto',
 			offset: '3vmin',
 			autogenerate: true,
 			tooltip: false,
@@ -279,15 +358,6 @@ const Plugin = () => {
 		options.indexbase = deck.getConfig()
 			.hashOneBasedIndex ? 1 : 0;
 
-		if (options.darktheme) {
-			if (!options.hasOwnProperty('color')) {
-				defaultOptions.color = 'white';
-			}
-			if (!options.hasOwnProperty('oppositecolor')) {
-				defaultOptions.oppositecolor = 'black';
-			}
-		}
-
 		defaults(options, defaultOptions);
 
 		function pluginPath() {
@@ -300,7 +370,6 @@ const Plugin = () => {
 			}
 			return path;
 		}
-
 		let VerticatorStylePath = options.csspath.verticator ? options.csspath.verticator : null || `${pluginPath()}verticator.css` || 'plugin/verticator/verticator.css'
 		let TooltipStylePath = options.csspath.tooltip ? options.csspath.tooltip : null  || `${pluginPath()}tooltip.css`  || 'plugin/verticator/tooltip.css'
 
@@ -309,7 +378,6 @@ const Plugin = () => {
 			console.log(`Verticator CSS path = ${VerticatorStylePath}`);
 			console.log(`Tooltip CSS path = ${TooltipStylePath}`);
 		}
-
 		loadStyle(VerticatorStylePath, 'stylesheet', function () {
 			if (options.tooltip) {
 				loadStyle(TooltipStylePath, 'stylesheet');